refactor: #28 Toast convention 적용
